Handle vswitch reconnect gracefully (vxlan + bridge fwd mode)
With vxlan bridge fwd mode, vxlanBridge will start the ovs switch object
in two of the vlan flooding objects (localFlood and allFlood). The switch object
is being populated when the first vxlan is assigned.

When user restart ovs switch, ofnet handle the reconnect which pass down the notification
to vxlanBridge object but it doesn't refresh the existing vlans to save the
new switch object. This causes netplugin stuck to talk to an unbuffered channel.

This patchset is to refresh vxlan entry to store new switch object
when ofnet handle ovs switch reconnection

Signed-off-by: Kahou Lei <kalei@cisco.com>